requirements.txt
boto3
botocore
faiss-cpu
fire
flake8
geocoder
geojson
geopandas
h3
hdbscan
invoke
ipykernel
joblib
jupyter
line_profiler
matplotlib
memory_profiler
metaflow
more_itertools
multiprocess
numpy
openai
pinecone-client
pip
plotly
plotnine
pyarrow
pydeck
pylint
pytest
pyyaml
qdrant-client
rasterio
rasterstats
requests
rtree
s3fs
scikit-image
scikit-learn
seaborn
snowflake-connector-python
sqlalchemy
tabula-py
tabulate
tenacity
toml
umap-learn
torch
torchvision
torchaudio
watchdog
wheel
xarray
yq

protobuf==3.20.*


